Purposefully explicitly limiting the cumulative_mdat_box_size to the 32-bit range. Given the use case scenario (mobile devices video transcoder where the size of video file is not reasonably expected to exceed 4GB limit) it makes perfect sense.

+++ cli/src/main.rs
@@ -21,7 +21,7 @@ struct Args {
     output: Option<PathBuf>,
 
     #[clap(long, short = 'c')]
-    cumulative_mdat_box_size: Option<u64>,
+    cumulative_mdat_box_size: Option<u32>,
 
     /// Path to the file to test sanitization on.
     file: PathBuf,
